+++ src/main/java/repository/road/DivideRoadRepository.java
@@ -1,38 +1,17 @@
 package repository.road;
 
 import config.JdbcTemplate;
+import domain.SqlReader;
 import java.sql.Connection;
 import java.sql.SQLException;
-import java.util.List;
-import domain.Shp;
-import domain.SqlReader;
 import repository.ExecuteQuery;
-import repository.road.SaveRoad;
 
-public class RoadRepository {
+public class DivideRoadRepository {
 
     private JdbcTemplate jdbcTemplate = new JdbcTemplate();
     private ExecuteQuery executeQuery = new ExecuteQuery();
 
-    public void saveOriginData(SqlReader createSql, List<Shp> shps) {
-        try (Connection conn = jdbcTemplate.getConnection()) {
-            createTable(conn, createSql);
-            saveRoad(conn, shps);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private void createTable(Connection conn, SqlReader sqlReader) {
-        executeQuery.create(conn, sqlReader);
-    }
-
-    private void saveRoad(Connection conn, List<Shp> shps) {
-        SaveRoad saveRoad = new SaveRoad();
-        saveRoad.save(conn, shps);
-    }
-
-    public void procOriginData(SqlReader createSql) {
+    public void run(SqlReader createSql) {
         try (Connection conn = jdbcTemplate.getConnection()) {
             createTable(conn, createSql);
             divideDumpedRoad(conn);
@@ -40,6 +19,10 @@ public class RoadRepository {
         } catch (SQLException e) {
             e.printStackTrace();
         }
+    }
+
+    private void createTable(Connection conn, SqlReader sqlReader) {
+        executeQuery.create(conn, sqlReader);
     }
 
     private void divideDumpedRoad(Connection conn) throws SQLException {
